# feat: enable receipt prefetching by default