@jakmeier 2022-09-21
@mm-near nominating you as a reviewer since you already reviewed the feature, this one is just the decision to activate it by default from now on